Added an AuthContainer file and fixed warnings
