+++ client/src/App.tsx
@@ -2,6 +2,7 @@ import React from "react";
 import "./App.css";
 
 import AuthContainer from "./containers";
+// import AuthContainer from "./components/AuthContainer";
 
 function App() {
   return (
